pyproject.toml
[project]
name = "mapas"
version = "0.1.0"
description = ""
authors = [
    {name = "Luciano Borges",email = "lgaubatz@unicamp.br"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit (>=1.51.0,<2.0.0)",
    "streamlit-folium (>=0.25.3,<0.26.0)",
    "folium (>=0.20.0,<0.21.0)",
    "prettymapp (>=0.5.0,<0.6.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "clipboard (>=0.0.4,<0.0.5)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
